requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

numpy>=1.24.3
opencv-python>=4.8.0.74
onnx>=1.14.0
insightface>=0.7.3
psutil>=5.9.5
tk>=0.1.0
customtkinter>=5.2.0
tkinterdnd2>=0.3.0; sys_platform != 'darwin' and platform_machine != 'arm64'
tkinterdnd2-universal>=1.7.3; sys_platform == 'darwin' and platform_machine == 'arm64'
pillow>=10.0.0
onnxruntime>=1.15.1; python_version != '3.9' and sys_platform == 'darwin' and platform_machine != 'arm64'
onnxruntime-coreml>=1.13.1; sys_platform == 'darwin' and platform_machine != 'arm64'
onnxruntime-silicon>=1.13.1; sys_platform == 'darwin' and platform_machine == 'arm64'
onnxruntime_gpu==1.17.0; sys_platform != 'darwin' 
tensorflow==2.15.0
protobuf>=4.23.4
tqdm>=4.65.0
gfpgan>=1.3.8
ffmpeg
